refactor: renamed  to transactionsProvider with references update. added slide from topedge to update in UI
