+++ lib/ui/split_group/split_group_list.dart
@@ -36,56 +36,63 @@ class SplitGroupsList extends ConsumerWidget {
     if (groups.isEmpty) {
       return const Center(child: Text('No groups'));
     }
-    return ListView.builder(
-      itemBuilder: (BuildContext context, int i) {
-        final groupId = groups[i].id;
-        final group = ref.watch(splitGroupProvider(groupId));
-        final participantsString = group.members
-            .where((u) => u.id != userId)
-            .map((u) => u.name)
-            .join(', ');
-        final bkgColor = colorFromString(group.id, 255, 0.7);
-        final tile = ListTile(
-          key: ValueKey(groupId),
-          title: Text(group.name),
-          leading: CircleAvatar(
-            backgroundColor: bkgColor,
-            child: Text(
-              group.name[0],
-            ),
-          ),
-          onTap: () {
-            navigateGroupHome(context, group);
-          },
-          subtitle: Wrap(
-            children: [
-              const Icon(
-                Icons.people,
-                size: 16,
-              ),
-              const SizedBox(width: 5),
-              Text(participantsString.isNotEmpty ? participantsString : 'You'),
-            ],
-          ),
-          trailing: IconButton(
-            icon: const Icon(
-              Icons.settings,
-              color: Colors.deepPurple,
-            ),
-            onPressed: () {
-              navigateGroupSettings(context, group);
-            },
-          ),
-        );
-        if (groupId == groupIdAnimateIn) {
-          return AnimatedTile(
-            duration: const Duration(milliseconds: 800),
-            child: tile,
-          );
-        }
-        return tile;
+    return RefreshIndicator.adaptive(
+      onRefresh: () async {
+        ref.invalidate(userSplitGroupsProvider);
       },
-      itemCount: groups.length,
+      child: ListView.builder(
+        itemBuilder: (BuildContext context, int i) {
+          final groupId = groups[i].id;
+          final group = ref.watch(splitGroupProvider(groupId));
+          final participantsString = group.members
+              .where((u) => u.id != userId)
+              .map((u) => u.name)
+              .join(', ');
+          final bkgColor = colorFromString(group.id, 255, 0.7);
+          final tile = ListTile(
+            key: ValueKey(groupId),
+            title: Text(group.name),
+            leading: CircleAvatar(
+              backgroundColor: bkgColor,
+              child: Text(
+                group.name[0],
+              ),
+            ),
+            onTap: () {
+              navigateGroupHome(context, group);
+            },
+            subtitle: Wrap(
+              children: [
+                const Icon(
+                  Icons.people,
+                  size: 16,
+                ),
+                const SizedBox(width: 5),
+                Text(
+                  participantsString.isNotEmpty ? participantsString : 'You',
+                ),
+              ],
+            ),
+            trailing: IconButton(
+              icon: const Icon(
+                Icons.settings,
+                color: Colors.deepPurple,
+              ),
+              onPressed: () {
+                navigateGroupSettings(context, group);
+              },
+            ),
+          );
+          if (groupId == groupIdAnimateIn) {
+            return AnimatedTile(
+              duration: const Duration(milliseconds: 800),
+              child: tile,
+            );
+          }
+          return tile;
+        },
+        itemCount: groups.length,
+      ),
     );
   }
