+++ lib/ui/transaction/expense_form/ui/new_expense_form.dart
@@ -23,7 +23,7 @@ class NewExpenseForm extends ConsumerWidget {
       ),
       body: BlocProvider(
         create: (context) => ExpenseFormBloc(
-          txNotifier: ref.read(transactionProvider(splitGroup.id).notifier),
+          txNotifier: ref.read(transactionsProvider(splitGroup.id).notifier),
           groupId: splitGroup.id,
           members: splitGroup.members,
           initialState: NewExpenseState(currency: splitGroup.defaultCurrency),
