+++ lib/ui/transaction/payment_form/ui/edit_payment_form.dart
@@ -45,7 +45,7 @@ class EditPaymentForm extends ConsumerWidget {
           members: splitGroup.members,
           currency: splitGroup.defaultCurrency,
           txNotifier: ref.read<TransactionNotifier>(
-              transactionProvider(splitGroup.id).notifier),
+              transactionsProvider(splitGroup.id).notifier),
           groupId: splitGroup.id,
           initialState: initialState,
         ),
